quality(build): unified strict warning policy + zero-warning build (Track B)
Warning policy harmonization:
- Add SECP256K1_WERROR CMake option (OFF default, -Werror/-WX)
- Add -Wconversion, -Wshadow, -Wformat=2, -Wundef globally
- security-audit.yml now uses -DSECP256K1_WERROR=ON (not raw CXX_FLAGS)
- OpenCL: remove duplicate global flags, keep MSVC-only suppressions
- STM32: add -Wextra, remove dangerous -Wno-return-type

Warning fixes (zero source warnings):
- glv.cpp: guard kMinusB1/B2/LambdaBytes with #ifndef __SIZEOF_INT128__
- ct_point.cpp: int -> size_t loop indices (sign-conversion)
- point.cpp: [[maybe_unused]] on scaffolding 4x64 functions,
  guard -Wrestrict pragma (GCC-only)

Test labels:
- Add 'core' label to all 13 core library tests (ctest -L core)

31/31 tests pass, zero source-level warnings.

+++ cpu/src/ct_point.cpp
@@ -1045,9 +1045,9 @@ static Scalar ct_scalar_mul_mod_n(const Scalar& a, const Scalar& b) noexcept {
 
     // --- 4x4 schoolbook multiply -> 8-limb product l[0..7] ---
     std::uint64_t l[8] = {};
-    for (int i = 0; i < 4; ++i) {
+    for (size_t i = 0; i < 4; ++i) {
         unsigned __int128 carry = 0;
-        for (int j = 0; j < 4; ++j) {
+        for (size_t j = 0; j < 4; ++j) {
             unsigned __int128 const t = (unsigned __int128)al[i] * bl[j] + l[i + j] + carry;
             l[i + j] = (std::uint64_t)t;
             carry = t >> 64;
